fix: resolve circular dependency in hypav3.ts
- Extract HypaV3Preset, HypaV3Settings types and createHypaV3Preset
  function to separate hypav3Preset.ts file
- Move HypaProcesserEx class inside hypaMemoryV3Main function to avoid
  module-level initialization error
- Update database.svelte.ts to import from hypav3Preset.ts

Fixes: "Cannot access 'HypaProcesser' before initialization" error
